refactor(mutations): better structured mutation rendering because of more modules.
