+++ apps/dashboard/src/components/mutations/mutationmodal/ModalMutation.vue
@@ -48,12 +48,13 @@ import type {
 } from '@sudosos/sudosos-client/src/api';
 import { useTransferStore } from '@/stores/transfer.store';
 import apiService from '@/services/ApiService';
-import TransactionDetailModal from '@/components/mutations/TransactionDetailModal.vue';
-import DepositDetailModal from '@/components/mutations/DepositDetailModal.vue';
-import InvoiceDetailModal from '@/components/mutations/InvoiceDetailModal.vue';
+import TransactionDetailModal from '@/components/mutations/mutationmodal/ModalDetailTransaction.vue';
+import DepositDetailModal from '@/components/mutations/mutationmodal/ModalDetailDeposit.vue';
+import InvoiceDetailModal from '@/components/mutations/mutationmodal/ModalDetailInvoice.vue';
+import FineDetailModal from '@/components/mutations/mutationmodal/ModalDetailFine.vue';
+import WaivedFineDetailModal from "@/components/mutations/mutationmodal/ModalDetailWaivedFine.vue";
 import router from '@/router';
 import { addListenerOnDialogueOverlay } from '@/utils/dialogUtil';
-import FineDetailModal from '@/components/mutations/FineDetailModal.vue';
 import { useI18n } from 'vue-i18n';
 import { useToast } from 'primevue/usetoast';
 import type { AxiosError } from 'axios';
@@ -61,7 +62,6 @@ import { handleError } from '@/utils/errorUtils';
 import { FinancialMutationType } from '@/utils/mutationUtils';
 import { UserRole } from '@/utils/rbacUtils';
 import { useAuthStore } from '@sudosos/sudosos-frontend-common';
-import WaivedFineDetailModal from "@/components/mutations/WaivedFineDetailModal.vue";
 import Skeleton from "primevue/skeleton";
 
 const props = defineProps({
