+++ apps/dashboard/src/modules/user/views/UserTransactionsView.vue
@@ -2,30 +2,18 @@
   <div class="page-container">
     <div class="page-title">{{ $t('transactions.Transactions') }}</div>
     <div class="content-wrapper gap-5 flex md:flex-column flex-column">
-      <MutationsTableComponent
+      <MutationsUserTabs
         class="transactions-table"
-        :header="$t('c_recentTransactionsTable.recent transactions')"
-        :paginatedMutationResponse="financialMutationsResponse"
-        :modal="true"
+        :get-balance-mutations="getUserMutations"
+        :get-seller-mutations="getTransactionsForOthers"
         :paginator="true"
-        :callbackFunction="getUserMutations"
-      />
-
-      <TransactionsTableComponent
-        class="transactions-table"
-        header="Transactions made for others"
-        :paginatedMutationResponse="financialMutationsResponse"
-        :modal="true"
-        :paginator="true"
-        :callbackFunction="getTransactionsForOthers"
       />
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
-import MutationsTableComponent from '@/components/mutations/MutationsTableComponent.vue';
-import TransactionsTableComponent from '@/components/mutations/TransactionsTableComponent.vue';
+import MutationsUserTabs from '@/components/mutations/MutationsUserTabs.vue';
 import apiService from '@/services/ApiService';
 import { useAuthStore, useUserStore } from '@sudosos/sudosos-frontend-common';
 import { ref } from 'vue';
